requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Assets==2.0
Flask-Dance==3.0.0
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.11.0
Flask-RESTful==0.3.8
flask-restx==0.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-sqlalchemy==0.22.3
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.15.7
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
URLObject==2.4.3
webassets==2.0
Werkzeug==0.16.1
WTForms==2.2.1
zipp==3.1.0


